requirements.txt
# Parsing
beautifulsoup4
lxml

# Rest
requests

# Formatting
black

# Testing
pytest

# SSH
paramiko[ed25519]
scp

# build
wheel
twine

# Images
Pillow
